% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.userterms-package.R
\docType{package}
\name{ergm.userterms-package}
\alias{ergm.userterms}
\alias{ergm.userterms-package}
\title{User-defined terms used in Exponential Family Random Graph Models}
\description{
This non-CRAN package contains template code for user defined (change) statistics
that can be used with the \CRANpkg{statnet} suite of packages (and \CRANpkg{ergm} in
particular). To use this package the \CRANpkg{statnet} packages \CRANpkg{ergm} and
\CRANpkg{network} are required.
}
\details{
As background, \CRANpkg{statnet} is a suite of software packages for statistical
network analysis. The packages implement recent advances in network modeling
based on exponential-family random graph models (ERGM). The components of
the package provide a comprehensive framework for ERGM-based network
modeling: tools for model estimation, for model evaluation, for model-based
network simulation, and for network visualization. This broad functionality
is powered by a central Markov chain Monte Carlo (MCMC) algorithm. The
coding is optimized for speed and robustness. For detailed information on
how to download and install the software, go to the \code{\link[ergm]{ergm}}
website: \url{statnet.org}. A tutorial, support newsgroup, references and
links to further resources are provided there.

We have invested a lot of time and effort in creating the \code{statnet}
suite of packages for use by other researchers. Please cite it in all papers
where it is used.

For complete citation information, use\cr
\code{citation(package="ergm")}.
}
\section{Creating User-defined terms using this package}{
 The \CRANpkg{statnet}
suite of packages allows the user to explore a large number of potential
models for their network data. These can be seen by typing
\code{?ergmTerm} (once the package \CRANpkg{ergm} has
been installed).  For more information on the models and terms see Morris,
Handcock, and Hunter (2008).

The purpose of the package is to allow additional terms can be coded up by
users (you!) and be used at native speeds with \CRANpkg{statnet}. So the suite
of packages can be extended with minimal work by the user. In addition the
core packages are not altered and so the new packages benefit from
improvements to the core suite.

The process of creating new terms is explained in depth by the
document entitled by Hunter, Goodreau, and Handcock (2013) that is
found in the \file{inst/doc} directory of this package. While that
article is still valid, API changes in \CRANpkg{ergm} 4.0 have
simplified the term implementations and added new capabilities; see
the \emph{Terms API} vignette in \CRANpkg{ergm}.

In brief, to add a new term you need to (all file references are relative to
the package directory):
\enumerate{
\item Download the source version of this package, currently hosted
at \url{https://github.com/statnet/ergm.userterms}, e.g.,
\url{https://github.com/statnet/ergm.userterms/archive/refs/heads/master.zip}.
\item Unpack the source into a directory.
\item Optionally, rename the package name from \code{ergm.userterms} to some
something more evocative of its use (e.g., \code{myergm}).  This can be done
using a global change to the files in the directory. Optionally, edit the
\file{DESCRIPTION} file to reflect the use of the package.
\item Edit \file{R/InitErgmTerm.users.R} to add \R \code{InitErgmTerm}
functions for the new terms.
\item Edit \file{src/changestats.users.c} to add \code{C} functions (like the
example already in that file) to compute the new change statistics.
\item Compile the package using the usual \R tools (e.g., \code{R CMD build
myergm} and \code{R CMD INSTALL myergm}).
\item Run it! It depends on \CRANpkg{ergm} and \CRANpkg{network}, of course.  See the
example below.
}
}

\references{
Hunter DR, Goodreau SM, Handcock MS (2013).  \code{ergm.userterms}: A
Template Package for Extending \code{statnet}, \emph{Journal of Statistical
Software} 52(2), 1-25, \doi{10.18637/jss.v052.i02}.
}
\seealso{
\link[ergm:ergm-package]{ergm-package}, \link[network:network-package]{network-package}, \code{\link[ergm:ergmTerm]{ergmTerm}}
}
\author{
\strong{Maintainer}: Pavel N. Krivitsky \email{pavel@statnet.org} (\href{https://orcid.org/0000-0002-9101-3362}{ORCID})

Authors:
\itemize{
  \item Mark S. Handcock \email{handcock@stat.ucla.edu}
  \item David R. Hunter \email{dhunter@stat.psu.edu}
  \item Carter T. Butts \email{buttsc@uci.edu}
  \item Steven M. Goodreau \email{goodreau@u.washington.edu}
  \item Martina Morris \email{morrism@u.washington.edu}
}

}
\keyword{models}
\keyword{package}
