% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.users.R
\name{mindegree-ergmTerm}
\alias{mindegree-ergmTerm}
\alias{InitErgmTerm.mindegree}
\title{Minimum Degree}
\usage{
# binary: mindegree(mindeg, by=NULL)
}
\arguments{
\item{mindeg}{an integer giving the}

\item{by}{an optional argument specifying a vertex attribute (see
Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for
details). If this is specified, then degrees are calculated using
the subnetwork consisting of only edges whose endpoints have the
same value of the \code{by} attribute.}
}
\description{
This term adds one network statistic to the model
being the number of nodes in the network of at least degree
\code{mindeg}. That is, the statistic equals the number of nodes
in the network with \code{mindeg} or more edges.  This term can
only be used with undirected networks.
}
\details{
This \code{InitErgmTerm} function is for the mindegree term
described by

Hunter DR, Goodreau SM, Handcock MS (2013).  \code{ergm.userterms}: A
Template Package for Extending \code{statnet}, \emph{Journal of Statistical
Software} 52(2), 1-25, \doi{10.18637/jss.v052.i02}.

It has been further updated to use the latest \CRANpkg{ergm} APIs
and conventions.
}
\seealso{
\code{\link[ergm:ergmTerm]{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "mindegree", "subsection")}
}
\concept{categorical nodal attribute}
\concept{undirected}
