% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.users.R
\name{sqrt.triangle-ergmTerm}
\alias{sqrt.triangle-ergmTerm}
\alias{InitErgmTerm.sqrt.triangle}
\alias{InitErgmTerm.sqrt.triangle.aux}
\alias{sqrt.triangle.aux-ergmTerm}
\title{Square root of the number of triangles}
\usage{
# binary: sqrt.triangle

# binary: sqrt.triangle.aux
}
\description{
These terms add one network statistic to the model,
the square root of the number of triangles. They demonstrate the
private and auxiliary storage mechanisms. The change statistic
for the square root of the number of triangles depends on the
initial number of triangles, so it is costly to compute -- unless
one can keep track of the number of triangles. These terms can
only be used with undirected networks.
}
\examples{

data(florentine)

sqrt(summary(flomarriage~triangle))
summary(flomarriage~sqrt.triangle)

stopifnot(sqrt(summary(flomarriage~triangle)) == summary(flomarriage~sqrt.triangle))

summary(flomarriage~sqrt.triangle.aux)

stopifnot(sqrt(summary(flomarriage~triangle)) == summary(flomarriage~sqrt.triangle.aux))
}
\seealso{
\code{\link[ergm:ergmTerm]{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "sqrt.triangle", "subsection")}
}
\concept{triadic}
\concept{undirected}
